Introduce secured endpoint
Using a free auth0 account, create a test api so that endpoints can be
secured.

This uses auth0 libraries to fetch JWKs and decode JWTs. This is done
via a standard http4k filter that allows to validate the token or
otherwise considered the request unauthorised.

To be able to test the security, a http client to fetch tokens is
introduced. It also uses the http4k to easily handle json request and
response in a type safe manner. The auth0 api client details are
expected as env variables so that they don't need to be hard-coded in
the tests.