Introduce secured endpoint expecting a token claim
Use bearer auth filter ability to capture token by refactoring jwt
verifier to decode tokens, introducing an abstraction to represent a
bearer token for our application so that it can be use to tighten
authorisation using a separate filter that expects a particular scope.

To test that an endpoint needs to include a specific token scope, create
a new client in the auth0 api that includes the token scope to show how
authorisation can be tighten further.